modify encoder

+++ tnet/channel.go
@@ -1,7 +1,7 @@
 package tnet
 
 import (
-	"context"
+	"errors"
 	"github.com/google/uuid"
 	"github.com/sisobobo/tinx/tlog"
 	"github.com/sisobobo/tinx/tpkg/bufio"
@@ -14,17 +14,16 @@ import (
 
 type Channel struct {
 	sync.RWMutex
-	id     string
-	server *Server
-	conn   net.Conn
-	rp, wp *bytes.Pool
-	rb, wb *bytes.Buffer
-	reader bufio.Reader
-	writer bufio.Writer
-	bucket *Bucket
-	ctx    context.Context
-	cancel context.CancelFunc
-	r      int
+	id       string
+	server   *Server
+	conn     net.Conn
+	rp, wp   *bytes.Pool
+	rb, wb   *bytes.Buffer
+	reader   bufio.Reader
+	writer   bufio.Writer
+	bucket   *Bucket
+	isClosed bool
+	r        int
 }
 
 func (c *Channel) Id() string {
@@ -39,26 +38,25 @@ func (c *Channel) LocalAddr() string {
 	return c.conn.LocalAddr().String()
 }
 
-func (c *Channel) SendMessage(message Message) {
-	c.RLock()
-	defer c.RUnlock()
+func (c *Channel) SendMessage(message Message) error {
 	data, err := c.server.codec.Encode(message)
 	if err != nil {
-		tlog.Errorf("encode error :", err)
-		return
+		return err
+	}
+	if c.isClosed {
+		return errors.New("connection is closed")
 	}
 	if _, err = c.writer.Write(data); err != nil {
-		tlog.Errorf("write error :", err)
-		return
+		return err
 	}
 	if err = c.writer.Flush(); err != nil {
-		tlog.Errorf("flush error :", err)
-		return
+		return err
 	}
+	return nil
 }
 
 func (c *Channel) Close() {
-	c.cancel()
+	c.close()
 }
 
 func NewChannel(server *Server, conn *net.TCPConn, r int) *Channel {
@@ -79,42 +77,35 @@ func NewChannel(server *Server, conn *net.TCPConn, r int) *Channel {
 }
 
 func (c *Channel) open() {
-	c.ctx, c.cancel = context.WithCancel(context.Background())
+	c.isClosed = false
 	c.server.handler.Connect(c)
-	go c.startReader()
-	select {
-	case <-c.ctx.Done():
-		c.close()
-		return
-	}
-}
-
-func (c *Channel) startReader() {
 	defer c.Close()
 	for {
-		select {
-		case <-c.ctx.Done():
+		msg, err := c.server.codec.Decode(&c.reader)
+		if c.isClosed {
 			return
-		default:
-			msg, err := c.server.codec.Decode(&c.reader)
-			if err == io.EOF {
-				break
-			}
-			if err != nil {
-				tlog.Errorf("decode error  :", err)
-				return
-			}
-			go c.server.handler.Receive(c, msg)
 		}
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			tlog.Errorf("decode error : %s", err)
+			return
+		}
+		c.server.handler.Receive(c, msg)
 	}
 }
 
 func (c *Channel) close() {
 	c.Lock()
 	defer c.Unlock()
+	if c.isClosed {
+		return
+	}
 	c.bucket.Remove(c)
 	c.rp.Put(c.rb)
 	c.wp.Put(c.wb)
 	c.conn.Close()
 	c.server.handler.DisConnect(c)
+	c.isClosed = true
 }
